+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Map
 
 * Fixed an extra large bottom margin of the following navigation camera.
-* Fixed a bug when the custom camera viewport was not applied. 
+* Fixed a bug when the custom camera viewport was not applied.
+* Fixed a bug with incorrect camera center when custom `GeometryFramingAfterManeuver.distanceToFrameAfterManeuver` is set.
 
 ### Other changes